+++ src/coff_parser.h
@@ -1,0 +1,48 @@
+/**
+ * Copyright Spack Project Developers. See COPYRIGHT file for details.
+ *
+ * SPDX-License-Identifier: (Apache-2.0 OR MIT)
+ */
+#pragma once
+
+#include <strsafe.h>
+#include <winnt.h>
+#include <string>
+
+#include "coff.h"
+#include "coff_reader_writer.h"
+
+class CoffParser {
+   private:
+    CoffReaderWriter* coffStream_;
+    coff coff_;
+    bool verified_ = false;
+    bool ParseData(PIMAGE_ARCHIVE_MEMBER_HEADER header, coff_member* member);
+    static void ParseShortImport(coff_member* member);
+    static void ParseFullImport(coff_member* member);
+    static void ParseFirstLinkerMember(coff_member* member);
+    static void ParseSecondLinkerMember(coff_member* member);
+    static void ReportLongImportMember(long_import_member* long_import);
+    static void ReportShortImportMember(short_import_member* short_import);
+    static void ReportLongName(const char* data);
+    void NormalizeLinkerMember(const std::string& name, const int& base_offset,
+                               const int& offset, const char* strings,
+                               DWORD symbols);
+    void NormalizeSectionNames(const std::string& name, char* section,
+                               const DWORD& section_data_start_offset,
+                               int data_size);
+    static bool ValidateLongName(coff_member* member, int size);
+    void writeRename(char* name, int size, int loc);
+    static bool matchesName(char* old_name, const std::string& new_name);
+
+   public:
+    explicit CoffParser(CoffReaderWriter* coff_reader);
+    ~CoffParser() = default;
+    bool Parse();
+    bool NormalizeName(std::string& name);
+    void Report();
+    int Verify();
+    static int Validate(std::string& coff);
+};
+
+bool reportCoff(CoffParser& coff);
